Display Bugfixes, Display-Ausgabe fuer RPi erweitert
- Bugfix RemoteCall-Display (fuer negative Paramterwerte)
- Bugfix Remotedisplay (ct-Sim) auf RPi
- Unterstuetzung von Display-Ausgabe auf Linux-Konsole fuer RPi (stdout oder externes Display)

+++ bot-logic/behaviour_remotecall.c
@@ -542,7 +542,7 @@ static uint8_t beh_selected = 0; /**< Status der Eingabe (0: nicht ausgewaehlt, 
  */
 static void keypad_param(char * data) {
 	if (keypad_param_index < REMOTE_CALL_MAX_PARAM && *data != 0) {
-		keypad_params[keypad_param_index].s16 = atoi(data);
+		keypad_params[keypad_param_index].s32 = atoi(data);
 		++keypad_param_index;
 	}
 	beh_selected = 1;
